% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{load_main_js}
\alias{load_main_js}
\title{Load util wizard scripts necessary for the package.}
\usage{
load_main_js()
}
\value{
A list of scripts.
}
\description{
This function loads the necessary utils scripts for the package.
}
\keyword{internal}
