% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{load_wizard}
\alias{load_wizard}
\title{Load scripts necessary for the package.}
\usage{
load_wizard()
}
\value{
A list of scripts.
}
\description{
This function loads the necessary scripts for the package.
}
\keyword{internal}
