% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{load_Wizard_css}
\alias{load_Wizard_css}
\title{Load css necessary for the package.}
\usage{
load_Wizard_css()
}
\value{
A list of scripts.
}
\description{
This function loads the necessary css for the package.
}
\keyword{internal}
