% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{load_wizard_js}
\alias{load_wizard_js}
\title{Load the Wizard-JS scripts necessary for the package.}
\usage{
load_wizard_js()
}
\value{
A list of scripts.
}
\description{
This function loads the necessary Wizard-JS js and css scripts for the package.
}
\keyword{internal}
