% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lock}
\alias{lock}
\title{Lock wizard

Server-side function to lock the wizard.}
\usage{
lock(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{wizard id}

\item{session}{shiny session}
}
\description{
Lock wizard

Server-side function to lock the wizard.
}
