% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reset}
\alias{reset}
\title{Reset wizard

Server-side function to reset the wizard.}
\usage{
reset(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{wizard id}

\item{session}{shiny session}
}
\description{
Reset wizard

Server-side function to reset the wizard.
}
