% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unlock}
\alias{unlock}
\title{Unlock wizard

Server-side function to unlock the wizard.}
\usage{
unlock(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{wizard id}

\item{session}{shiny session}
}
\description{
Unlock wizard

Server-side function to unlock the wizard.
}
