% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizard.R
\name{wizard}
\alias{wizard}
\title{start wizard}
\usage{
wizard(
  orientation = c("horizontal", "vertical"),
  style = c("dots", "tabs", "progress"),
  show_buttons = c(TRUE, FALSE),
  ...
)
}
\value{
wizard html
}
\description{
start wizard
}
