% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{wizard_dependencies}
\alias{wizard_dependencies}
\title{Load util wizard scripts necessary for the package.}
\usage{
wizard_dependencies()
}
\value{
A list of scripts.
}
\description{
This function loads the necessary utils scripts for the package.
}
\keyword{internal}
