% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizard.R
\name{wizard_show}
\alias{wizard_show}
\title{Show wizard}
\usage{
wizard_show(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{wizard id}

\item{session}{shiny session}
}
\description{
Show wizard
}
