% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizard.R
\name{wizard_step}
\alias{wizard_step}
\title{Add a step to the wizard}
\usage{
wizard_step(
  ...,
  step_title = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{...}{step content}

\item{step_title}{Title of the step (if it will be Step 1, Step 2, etc.)}

\item{session}{shiny session}
}
\value{
wizard step html
}
\description{
Add a step to the wizard
}
