% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizard.R
\name{wizard}
\alias{wizard}
\title{start wizard component}
\usage{
wizard(
  ...,
  orientation = "horizontal",
  style = "progress",
  show_buttons = TRUE,
  id = NULL,
  modal = TRUE,
  options = list()
)
}
\arguments{
\item{...}{wizard content}

\item{orientation}{wizard orientation (horizontal or vertical)}

\item{style}{wizard style (dots, tabs or progress)}

\item{show_buttons}{show buttons or not (TRUE or FALSE)}

\item{id}{wizard id}

\item{modal}{modal or not (TRUE or FALSE)}

\item{options}{A list of options. See the documentation of
'Wizard-JS' (<URL: https://github.com/AdrianVillamayor/Wizard-JS>) for
possible options.}
}
\value{
wizard html
}
\description{
This file contains the implementation of the wizard functionality in R.
The wizard module provides various functions and utilities for performing magical operations.
It includes functions for casting spells, brewing potions, and summoning mystical creatures.
Use this module to unleash your inner wizard and explore the world of R magic!
}
